dev-requirements.txt
apipkg==1.5
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
colorama==0.4.1
Django==2.2
django-filter==2.1.0
djangorestframework==3.9.2
execnet==1.6.0
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==7.0.0
mysqlclient==1.4.2.post1
pep8==1.7.1
pluggy==0.9.0
py==1.8.0
pylint==2.3.1
pylint-django==2.0.8
pylint-plugin-utils==0.5
pytest==4.4.1
pytest-cache==1.0
pytest-pep8==1.0.6
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.4
wrapt==1.11.1
yapf==0.27.0
